basic row validation, add alert on csv error

+++ admin/kupons.php
@@ -30,20 +30,37 @@ $coupon      = null;
 $alertHelper = Shop::Container()->getAlertService();
 $errors      = [];
 $res         = handleCsvImportAction('kupon', static function ($obj, &$importDeleteDone, $importType = 2) {
-    $db = Shop::Container()->getDB();
-    if ($importType === 0 && $importDeleteDone === false) {
-        $db->query('TRUNCATE TABLE tkupon', ReturnType::AFFECTED_ROWS);
-        $db->query('TRUNCATE TABLE tkuponsprache', ReturnType::AFFECTED_ROWS);
-        $importDeleteDone = true;
-    }
-
+    $db          = Shop::Container()->getDB();
     $couponNames = [];
+    $cols        = Shop::Container()->getDB()->queryPrepared(
+        'SELECT `column_name` AS name
+            FROM information_schema.columns 
+            WHERE `table_schema` = :sma
+                AND `table_name` = :tn',
+        ['sma' => DB_NAME, 'tn' => 'tkupon'],
+        ReturnType::COLLECTION
+    )->map(function ($e) {
+        return $e->name;
+    })->toArray();
 
     foreach (get_object_vars($obj) as $key => $val) {
         if (mb_strpos($key, 'cName_') === 0) {
             $couponNames[mb_substr($key, 6)] = $val;
             unset($obj->$key);
         }
+        if (!in_array($key, $cols, true)) {
+            unset($obj->$key);
+        }
+    }
+    if (!isset($obj->cCode) || !isset($obj->nGanzenWKRabattieren) || !isset($obj->cKunden) || !isset($obj->cKategorien)
+        || !isset($obj->cHersteller) || !isset($obj->cArtikel)
+    ) {
+        return false;
+    }
+    if ($importType === 0 && $importDeleteDone === false) {
+        $db->query('TRUNCATE TABLE tkupon', ReturnType::AFFECTED_ROWS);
+        $db->query('TRUNCATE TABLE tkuponsprache', ReturnType::AFFECTED_ROWS);
+        $importDeleteDone = true;
     }
 
     if (isset($obj->cCode) && $db->select('tkupon', 'cCode', $obj->cCode) !== null) {
@@ -67,8 +84,12 @@ $res         = handleCsvImportAction('kupon', static function ($obj, &$importDel
 }, [], null, 2, $errors);
 
 if ($res > 0) {
-    foreach ($errors as $key => $error) {
-        $alertHelper->addAlert(Alert::TYPE_ERROR, $error, 'errorImportCSV_' . $key);
+    if (count($errors) > 0) {
+        foreach ($errors as $key => $error) {
+            $alertHelper->addAlert(Alert::TYPE_ERROR, $error, 'errorImportCSV_' . $key);
+        }
+    } else {
+        $alertHelper->addAlert(Alert::TYPE_ERROR, __('errorImportCSV'), 'errorImportCSV');
     }
 } elseif ($res === 0) {
     $alertHelper->addAlert(Alert::TYPE_SUCCESS, __('successImportCSV'), 'successImportCSV');
@@ -278,20 +299,20 @@ if ($action === 'bearbeiten') {
     );
 
     $smarty->assign('tab', $tab)
-           ->assign('oFilterStandard', $filterStandard)
-           ->assign('oFilterVersand', $filterVersand)
-           ->assign('oFilterNeukunden', $filterNeukunden)
-           ->assign('oPaginationStandard', $paginationStandard)
-           ->assign('oPaginationVersandkupon', $paginationVersand)
-           ->assign('oPaginationNeukundenkupon', $paginationNeukunden)
-           ->assign('oKuponStandard_arr', $standardCoupons)
-           ->assign('oKuponVersandkupon_arr', $shippingCoupons)
-           ->assign('oKuponNeukundenkupon_arr', $newCustomerCoupons)
-           ->assign('nKuponStandardCount', $nKuponStandardTotal)
-           ->assign('nKuponVersandCount', $nKuponVersandTotal)
-           ->assign('nKuponNeukundenCount', $nKuponNeukundenTotal);
+        ->assign('oFilterStandard', $filterStandard)
+        ->assign('oFilterVersand', $filterVersand)
+        ->assign('oFilterNeukunden', $filterNeukunden)
+        ->assign('oPaginationStandard', $paginationStandard)
+        ->assign('oPaginationVersandkupon', $paginationVersand)
+        ->assign('oPaginationNeukundenkupon', $paginationNeukunden)
+        ->assign('oKuponStandard_arr', $standardCoupons)
+        ->assign('oKuponVersandkupon_arr', $shippingCoupons)
+        ->assign('oKuponNeukundenkupon_arr', $newCustomerCoupons)
+        ->assign('nKuponStandardCount', $nKuponStandardTotal)
+        ->assign('nKuponVersandCount', $nKuponVersandTotal)
+        ->assign('nKuponNeukundenCount', $nKuponNeukundenTotal);
 }
 
 $smarty->assign('action', $action)
-       ->assign('couponTypes', Kupon::getCouponTypes())
-       ->display('kupons.tpl');
+    ->assign('couponTypes', Kupon::getCouponTypes())
+    ->display('kupons.tpl');
